log_sentry file rename
